fix: move order grouping to utils
