+++ client-app/shared/cart/composables/useCart.ts
@@ -37,9 +37,9 @@ import { useNotifications } from "@/shared/notification";
 import {
   ClearCartModal,
   ExtendedGiftItemType,
+  getLineItemsGroupedByVendor,
   getLineItemValidationErrorsGroupedBySKU,
   OutputBulkItemType,
-  TGroupedItems,
   TGroupItem,
 } from "@/shared/cart";
 
@@ -52,33 +52,7 @@ const payment = computed<PaymentType | undefined>(() => cart.value.payments?.[0]
 const availableShippingMethods = computed<ShippingMethodType[]>(() => cart.value.availableShippingMethods ?? []);
 const availablePaymentMethods = computed<PaymentMethodType[]>(() => cart.value.availablePaymentMethods ?? []);
 
-const lineItemsGroupedByVendor = computed<TGroupItem[]>(() => {
-  // NOTE: The group without the vendor should be displayed last.
-  const groupWithoutVendor: TGroupItem = { items: [] };
-  const map: TGroupedItems = {};
-
-  cart.value.items?.forEach((item) => {
-    const vendor = item.product?.vendor;
-
-    if (vendor) {
-      const vendorId = vendor.id;
-
-      map[vendorId] = map[vendorId] || { vendor, items: [] };
-      map[vendorId].items.push(item);
-    } else {
-      groupWithoutVendor.items.push(item);
-    }
-  });
-
-  const result = Object.values(map)
-    // Sort by Vendor
-    .sort((a, b) => a.vendor!.name.localeCompare(b.vendor!.name));
-
-  // Add the group without the vendor to the end.
-  result.push(groupWithoutVendor);
-
-  return result;
-});
+const lineItemsGroupedByVendor = computed<TGroupItem[]>(() => getLineItemsGroupedByVendor(cart.value.items || []));
 
 const addedGiftsByIds = computed(() => keyBy(cart.value.gifts, "id"));
