+++ client-app/shared/cart/utils/index.ts
@@ -1,5 +1,5 @@
-import { OrderLineItemType, ValidationErrorType } from "@/xapi";
-import { ItemForAddBulkItemsToCartResultsPopup, OutputBulkItemType } from "@/shared/cart";
+import { LineItemType, OrderLineItemType, ValidationErrorType } from "@/xapi";
+import { ItemForAddBulkItemsToCartResultsPopup, OutputBulkItemType, TGroupedItems, TGroupItem } from "@/shared/cart";
 
 export function getLineItemValidationErrorsGroupedBySKU(
   errors: ValidationErrorType[] = []
@@ -34,4 +34,32 @@ export function getItemsForAddBulkItemsToCartResultsPopup(
     slug: item.product?.slug,
     errors: errorsGroupedBySKU[item.sku!],
   }));
+}
+
+export function getLineItemsGroupedByVendor(items: LineItemType[]): TGroupItem[] {
+  // NOTE: The group without the vendor should be displayed last.
+  const groupWithoutVendor: TGroupItem = { items: [] };
+  const map: TGroupedItems = {};
+
+  items.forEach((item) => {
+    const vendor = item.product?.vendor;
+
+    if (vendor) {
+      const vendorId = vendor.id;
+
+      map[vendorId] = map[vendorId] || { vendor, items: [] };
+      map[vendorId].items.push(item);
+    } else {
+      groupWithoutVendor.items.push(item);
+    }
+  });
+
+  const result = Object.values(map)
+    // Sort by Vendor
+    .sort((a, b) => a.vendor!.name.localeCompare(b.vendor!.name));
+
+  // Add the group without the vendor to the end.
+  result.push(groupWithoutVendor);
+
+  return result;
 }
